refactor: SwipeCarousel 및 하위 컴포넌트 분리
